+++ tests/testthat/test-crew_controller.R
@@ -428,3 +428,10 @@ crew_test("backlog with saturation", {
     x$push(Sys.sleep(30), scale = FALSE)
   }
 })
+
+crew_test("descale", {
+  controller <- crew_controller_local()
+  expect_null(controller$autoscaling)
+  controller$descale()
+  expect_false(controller$autoscaling)
+})